Remove the ability to mix free and paid pages (#658)
There is a quite annoying bug where if a user prints using free pages, their free page pool is only decreased by p and not n*p (where n is the number of copies, and p is the number of pages in the document).
The quickest and easiest way of dealing with it is to only allow a print job to be fully paid or entirely paid and calculate the printed pages correctly.
This bug is quite annoying as our registry shows only a fraction of the actual usage (i.e. 5000 pages of plain pages was loaded into the printer, and our Mars instance shows only a few printed pages).